# public/openapi.yaml
openapi: 3.0.3
info:
  title: ConejoMalo API
  version: '1.0.0'
  description: |
    API pública y de administración para la tienda. Autenticación vía cookie `session`.
    Este documento es la fuente de verdad de los payloads.
servers:
  - url: https://{host}/api
    variables:
      host:
        default: localhost:3000
  - url: http://localhost:3000/api
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string, nullable: true }
        fields:
          type: object
          additionalProperties: { type: string }
          nullable: true
...
    UserListItem:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        roles:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        price: { oneOf: [ { type: string }, { type: number } ] }
        imageUrl: { type: string, nullable: true }
        categoryId: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        category:
          type: object
          nullable: true
          properties:
            id: { type: string }
            name: { type: string }
            slug: { type: string }
    ProductAdminLookup:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        price: { oneOf: [ { type: string }, { type: number } ] }
        category:
          type: object
          nullable: true
          properties:
            id: { type: string }
            name: { type: string }
...
paths:
...
  /admin/products:
    get:
      summary: Listar productos (admin)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: sort
          schema: { type: string, enum: [new, price_asc, price_desc, name] }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: cat
          schema: { type: string, description: 'ID o slug de la categoría' }
        - in: query
          name: active
          schema: { type: string, enum: ["true","false","all"], default: 'all' }
      responses:
        '200':
          description: Lista o paginación (si hay parámetros)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/Product' }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Product' }
                      page: { type: integer }
                      per: { type: integer }
                      total: { type: integer }
                      totalPages: { type: integer }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
    post:
      summary: Crear producto (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, categoryId]
              properties:
                name: { type: string }
                price: { oneOf: [ { type: number }, { type: string } ] }
                description: { type: string, nullable: true }
                categoryId: { type: string }
                active: { type: boolean }
          multipart/form-data:
            schema:
              type: object
              required: [name, price, categoryId]
              properties:
                name: { type: string }
                price: { type: string }
                description: { type: string }
                categoryId: { type: string }
                active: { type: string, enum: ["true","false"] }
                image:
                  type: string
                  format: binary
      responses:
        '201': { description: Creado, content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '400': { description: Error de validación, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
...
  /admin/products/search:
    get:
      summary: Búsqueda rápida de productos (admin)
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: cat
          schema: { type: string, description: 'ID o slug de la categoría' }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ProductAdminLookup' } } } } }
        '401': { description: No autenticado }
        '403': { description: Prohibido }

  # Admin: Categories
...
  /admin/users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Obtener usuario (admin)
      responses:
        '200': { description: Ok, content: { application/json: { schema: { $ref: '#/components/schemas/UserListItem' } } } }
        '404': { description: No encontrado }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
    put:
      summary: Actualizar usuario (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                name: { type: string, nullable: true }
                roles: { type: array, items: { type: string } }
                password: { type: string }
      responses:
        '200': { description: Actualizado, content: { application/json: { schema: { $ref: '#/components/schemas/UserListItem' } } } }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
    delete:
      summary: Eliminar usuario (admin)
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
...
  /admin/roles/{id}/users:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Listar usuarios de un rol (admin)
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/UserListItem' } } } } } } }
        '404': { description: No encontrado }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
  /admin/roles/{id}/users/{userId}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
      - in: path
        name: userId
        required: true
        schema: { type: string }
    delete:
      summary: Revocar rol a usuario (admin)
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '404': { description: No encontrado }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
        '404': { description: No encontrado }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
    delete:
      summary: Revocar rol a usuario (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string }
      responses:
        '200': { description: Ok, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '404': { description: No encontrado }
        '401': { description: No autenticado }
        '403': { description: Prohibido }
...
security:
  - cookieAuth: []